requirements.txt
anthropic
pillow
plotly
kaleido
python-dotenv
requests
